refactor(solver): split application evaluation into its own module (#13002)
`evaluation/evaluate.rs` had grown to 2002 lines, over the repo's 2000-line
hard cap, mixing the core evaluator state and `evaluate` dispatch loop with
the large `evaluate_application` orchestrator family.

Extract the application-evaluation machinery (callee normalization, per-DefId
recursion/divergence guards, context building, known-/extracted-params body
dispatch, homomorphic mapped passthrough and union distribution, the
application-eval cache gates, and display-alias bookkeeping) verbatim into a
new sibling module `evaluation/evaluate/application.rs`, following the existing
`use super::*;` + `impl TypeEvaluator` split idiom used by support.rs,
closed_eval.rs, and display_alias.rs.

No semantic change: method bodies are unchanged. `evaluate_application`,
`resolve_application_def_id`, and `store_intermediate_application_display_alias`
are widened from private to `pub(super)` so their existing cross-module and
test callers continue to resolve once the methods live in a sibling module.

evaluate.rs drops to 804 lines; application.rs is 1216 lines. Both are
comfortably under the cap.
